fix(example-app): :bug: fix warnings refactor code add icons
