+++ example/src/components/index.ts
@@ -1,1 +1,3 @@
+export * from "./Background";
+export * from "./EasingLogo";
 export * from "./Group";
