+++ example/src/pages/AboutScreen.tsx
@@ -1,104 +1,23 @@
-import React, { useMemo, useRef } from "react";
-import { Animated, Easing, Image, Linking } from "react-native";
+import React from "react";
+import { Linking } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import { Path, Pattern, Rect, Svg } from "react-native-svg";
-import { Box, Button, Text, useTheme } from "@adaptui/react-native-tailwind";
+import {
+  Box,
+  Button,
+  Dribble,
+  Github,
+  Icon,
+  Text,
+  useTheme,
+} from "@adaptui/react-native-tailwind";
 import Constants from "expo-constants";
 
-import { Group } from "../components";
+import Background from "../components/Background";
+import EasingLogo from "../components/EasingLogo";
 
 const AboutScreen = ({ navigation }) => {
   const { bottom } = useSafeAreaInsets();
   const tailwind = useTheme();
-
-  function useMountEffect(effect: React.EffectCallback): void {
-    React.useEffect(() => {
-      effect();
-    }, []); // eslint-disable-line
-  }
-
-  const Background = props => {
-    const { color, stroke, gap, transform } = props;
-    const { x, y, k } = transform || { x: 0, y: 0, k: 1 };
-    const scaledGap = (gap || 15) * k;
-    const patternId = useMemo(
-      () => `patten-${Math.floor(Math.random() * 100000)}`,
-      [],
-    );
-
-    return (
-      <Svg
-        style={[tailwind.style("w-full h-full absolute"), { zIndex: -100 }]}
-        width={"100%"}
-        height={"100%"}
-      >
-        <Pattern
-          id={patternId}
-          x={x % scaledGap}
-          y={y % scaledGap}
-          width={scaledGap}
-          height={scaledGap}
-          patternUnits={"userSpaceOnUse"}
-        >
-          <Path
-            stroke={color || "#dddddd"}
-            strokeWidth={stroke || 1}
-            d={`M${scaledGap / 2} 0 V${scaledGap} M0 ${
-              scaledGap / 2
-            } H${scaledGap}`}
-          />
-        </Pattern>
-        <Rect
-          x={0}
-          y={0}
-          width={"100%"}
-          height={"100%"}
-          fill={`url(#${patternId})`}
-        />
-      </Svg>
-    );
-  };
-
-  const EasingLogo = () => {
-    const easeValue = useRef(new Animated.Value(0));
-
-    useMountEffect(() => {
-      Animated.loop(
-        Animated.sequence([
-          Animated.timing(easeValue.current, {
-            toValue: 10,
-            duration: 1000,
-            easing: Easing.bezier(0.56, 0.15, 0.45, 0.87),
-            useNativeDriver: true,
-          }),
-          Animated.timing(easeValue.current, {
-            toValue: 0,
-            duration: 1000,
-            easing: Easing.bezier(0.56, 0.15, 0.45, 0.87),
-            useNativeDriver: true,
-          }),
-        ]),
-      ).start();
-    });
-
-    const ease = easeValue.current.interpolate({
-      inputRange: [0, 10],
-      outputRange: [0, 5],
-    });
-
-    return (
-      <Animated.View
-        style={{
-          transform: [{ translateY: ease }],
-        }}
-      >
-        <Image
-          source={require("../../assets/logo.png")}
-          style={tailwind.style("h-50 w-50")}
-        />
-      </Animated.View>
-    );
-  };
 
   return (
     <Box style={tailwind.style("flex-1")}>
@@ -120,45 +39,56 @@ const AboutScreen = ({ navigation }) => {
           </Text>
         </Box>
 
-        <Group label="" style={tailwind.style("mt-auto shadow-xl")}>
-          <Box style={tailwind.style("flex-col w-full p-2")}>
+        <Box
+          style={tailwind.style(
+            "flex-row flex-wrap justify-center items-center mt-auto",
+          )}
+        >
+          <Box style={tailwind.style("justify-center items-center w-full")}>
             <Button
               onPress={() => {
                 navigation.navigate("Components");
               }}
-              size="lg"
+              size="xl"
             >
               View Components
             </Button>
+          </Box>
+
+          <Box
+            style={tailwind.style(
+              " flex-row justify-center items-center w-full mt-4",
+            )}
+          >
             <Button
+              prefix={
+                <Icon
+                  icon={<Github fill={tailwind.getColor("text-white-500")} />}
+                />
+              }
               onPress={() => {
-                Linking.openURL("https://timeless.co");
+                Linking.openURL("https://github.com/timelessco");
               }}
-              size="lg"
-              style={tailwind.style("mt-6")}
+              size="xl"
             >
-              Visit our website
+              Github
             </Button>
             <Button
+              prefix={
+                <Icon
+                  icon={<Dribble fill={tailwind.getColor("text-white-500")} />}
+                />
+              }
               onPress={() => {
-                Linking.openURL("https://twitter.com/timelessco");
+                Linking.openURL("https://dribbble.com/timelessco");
               }}
-              size="lg"
-              style={tailwind.style("mt-6")}
+              size="xl"
+              style={tailwind.style("ml-2")}
             >
-              Follow us on Twitter
-            </Button>
-            <Button
-              onPress={() => {
-                Linking.openURL("mailto:hello@timeless.co");
-              }}
-              size="lg"
-              style={tailwind.style("mt-6")}
-            >
-              Contact Us
+              Dribble
             </Button>
           </Box>
-        </Group>
+        </Box>
         <Text style={tailwind.style("font-medium pt-2")}>
           version: {Constants.manifest.version}(5)
         </Text>
