+++ src/icons/index.ts
@@ -12,11 +12,13 @@ export * from "./CloseSmall";
 export * from "./Dash";
 export * from "./DefaultUser";
 export * from "./Delete";
+export * from "./Dribble";
 export * from "./Equals";
 export * from "./Error";
 export * from "./FastBackward";
 export * from "./FastForward";
 export * from "./Filter";
+export * from "./Github";
 export * from "./Image";
 export * from "./Info";
 export * from "./Link";
@@ -26,4 +28,5 @@ export * from "./Search";
 export * from "./Slash";
 export * from "./Slot";
 export * from "./Slot";
+export * from "./Timeless";
 export * from "./UpDownArrow";
